new: decrease autoCommit time to reduce Java Memory
This isn't an exact science and there are a lot of variables.  The general idea is to commit documents to disk so that Java Memory can be more appropriately managed by Solr and not run out of it.  If an solr index is going at 10 docs/sec, it will index 50 docs in 5 secs, if it's only going at 5 docs/sec, it will index 25 docs in 5 secs.  Therefore, there isn't a good method for identifying how many documents will be indexed within a timeframe nor is it guaranteed to know the size of those documents.  For our CKAN application, I'm more okay with this number, but it can be debated and retheorized.